Added include order to  README.md. changed all floats to doubles to avoid truncation warnings.

+++ src/main/cpp/Odometry/Limelight.cpp
@@ -14,7 +14,7 @@ bool Limelight::HasTarget()
   return table->GetBoolean("tv", false);
 }
 
-float Limelight::xOffset()
+double Limelight::xOffset()
 {
   return table->GetNumber("tx", 0);
 }
